@@ -1,36 +1,28 @@
-import React, {Component} from 'react';
+import React, {Component, useState, useEffect} from 'react';
 import logo from './logo.svg';
 import './App.css';
 
-class App extends Component {
+function App () {
+    const [message, setMessage] = useState("");
 
-    state = {};
-
-    componentDidMount() {
-        setInterval(this.hello, 250);
-    }
-
-    hello = () => {
+    useEffect(() => {
         fetch('/api/hello')
             .then(response => response.text())
             .then(message => {
-                this.setState({message: message});
+                setMessage(message);
             });
-    };
-
-    render() {
-        return (
-            <div className="App">
-                <header className="App-header">
-                    <img src={logo} className="App-logo" alt="logo"/>
-                    <h1 className="App-title">{this.state.message}</h1>
-                </header>
-                <p className="App-intro">
-                    To get started, edit <code>src/App.js</code> and save to reload.
-                </p>
-            </div>
-        );
-    }
+    },[])
+    return (
+        <div className="App">
+        <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo"/>
+        <h1 className="App-title">{message}</h1>
+        </header>
+        <p className="App-intro">
+        To get started, edit <code>src/App.js</code> and save to reload.
+    </p>
+    </div>
+)
 }
 
 export default App;
